+++ ssrf/__init__.py
@@ -1,0 +1,1 @@
+# SSRF detection package